requirements.txt
python-dotenv
PyPDF2
transformers
sentence-transformers
scikit-learn
langchain
langchain-groq
streamlit
numpy
tensorflow
